Extract pager parsing logic into shared utility
